requirements.txt
pytest>=7.1.3
requests>=2.25.0
PyMySQL>=0.10.1
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0